examples/confluent-example/pyproject.toml
[tool.poetry]
name = "confluent-example"
version = "0.1.0"
description = ""
authors = ["Marcos Schroh <marcos.schroh@kpn.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kstreams = { path = "../../.", develop = true }
python-schema-registry-client = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "confluent_example.app:main"
